requirements.txt
# pip install -r requirements.txt
flask
flask-cors
google-cloud-storage
pymongo
dnspython
